issue with hardware means that screen, MCP and SD cannot be ran off the same pico, to get around this a headless, uart over usb and uart to pico versions are being made, this is the first successful test of the uart over usb version

+++ UICode/pid.h
@@ -11,7 +11,7 @@ typedef struct PIDEntry {
     char *name;
     char *unit;
     PIDForm form;
-    int divisor;
+    float divisor;
     bool supported;
 } PIDEntry;
